Support for gpio connected ir senders/receivers
This module allows users to use a GPIO connected
IR receiver and/or sender with LIRC. The module
uses the sysfs gpio kernel functions to interact
with the gpio. You can configure the module from
inside the device tree like this:

lirc_gpio {
	compatible = "lirc_gpio";
	gpios = <&gpio3 6 1 &gpio3 7 2>;
	pinctrl-names = "default";
	pinctrl-0 = <&pinctrl_hummingboard_gpio3_6>;
	pinctrl-1 = <&pinctrl_hummingboard_gpio3_7>;
	linux,sense = <-1>;
	linux,softcarrier = <1>;
	linux,validgpios = <1 73 72 71 70 194 195 67>;
};

It should therefor work on all platforms.