dts: microsom-ar8035: fix enet reference pll
Commit d24a5df30aed6e85e19e9bfe92f184b933debbdb changed the
default clock of enet_ref_pll to 125Mhz which breaks our ethernet.
Since it is only one board that needs this it would make sense to
revert that commit and just patch that single device-tree.  Since
This is already accepted I will patch our device-tree to work
around this commit.