mach: imx: Re-enable SmartEEE in the Atheros Phy
I am testing the atheros phy's built in power savings features,
and did find that just enabling them was causing link disconnects
as reported initially.  However since increasing the tx twd time
out I have not experienced any more disconnects.

This is a feature that I want to get working as it saves about 60mA's
of power when the network is idle.