regulator: core: Make the state sysfs property r/w
There are many instances where it is beneficial for userspace
to be able to fully power off a device, be it powersaving, device
recovery, or testing.  All regulators will support the enabled/
disabled state so make this core property r/w and hook it into
the core enable/disable functions.