net: fec: fix the no_egiga support
Boundary added a patch to allow disabling gigabit ethernet support
in order to help with compatibility with switches that do not have
pause frames enabled.  This worked for the initial connectivity
but didn't fully disable support in the driver so reconnections would
come back as gigabit.  This patch fixes this and adds the expected
behaviour of no 1000BaseT support.

Signed-off-by: Jon Nettleton <jon@solid-run.com>